+++ lib/SensorFusion/src/SensorFusion.h
@@ -43,7 +43,7 @@
  * Includes
  *****************************************************************************/
 
-#include "SensorData.h"
+#include "SerialMuxChannels.h"
 #include "LinearKalmanFilter.h"
 #include <stdint.h>
 #include <SensorConstants.h>
@@ -100,7 +100,7 @@ private:
      */
     void transfromLocalToGlobal(int16_t * globalResult, const int16_t * localVectorToTransform, const int16_t & rotationAngle);
 
-    void estimateAngle(int16_t & estimatedAngle, const int16_t & encoderAngle, const int16_t & magnetometerValueX, const int16_t & magnetometerValueY);
+    void estimateAngle(int16_t & estimatedAngle, const int32_t & encoderAngle, const int16_t & magnetometerValueX, const int16_t & magnetometerValueY);
 };
 
 /******************************************************************************
